Updates to get_by_conditions
Added optional keys for the request json that if "fuzzy" is specified and is true it will make it a fuzzy search, if the gt_lt key exists it must be a string of either "<" or ">" and all the conditions must be values that can be compared using > or < and all the conditions will be checked by this for example if your conditions are price and quality both having the condition larger than 50 using gt_lt: ">" then both will have to be larger than 50. If both fuzzy and gt_lt are set then fuzzy takes preference and gt_lt are excluded. Both gt_lt and fuzzy dont need to be set though.